requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
jinja2==3.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
qrcode[pil]==7.4.2
aiosmtplib==3.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
itsdangerous==2.1.2
